Create README.md for GitHub profile
This commit introduces the README.md file to my GitHub profile. The README includes a detailed overview of my professional background, key technologies I work with, and highlight projects. It serves as an introduction to collaborators, potential employers, and anyone interested in my work. This file will help in making my GitHub more personable and provide a quick snapshot of my skills and projects.